Deleted BaseLayout, ProjectLayout and ContractorLayout restored separately, deleted menu switch in AppSidebar, created ContractorTopbar
