+++ src/layout/AppSidebar.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import ContractorMenu from './ContractorMenu.vue';
 import ManagerMenu from './ManagerMenu.vue';
 import TenancyMenu from './TenancyMenu.vue';
 import { useProjectStore } from '@/stores/ProjectStore';
@@ -16,15 +15,6 @@ const projectStore = useProjectStore();
     class="layout-sidebar"
   >
     <ManagerMenu></ManagerMenu>
-  </div>
-  <div
-    v-if="
-      projectStore.selectedProject?.memberRole === 'CONSULTANT' ||
-      projectStore.selectedProject?.memberRole === 'CARETAKER'
-    "
-    class="layout-sidebar"
-  >
-    <ContractorMenu></ContractorMenu>
   </div>
   <div v-if="projectStore.selectedProject?.memberRole === 'LESSEE'" class="layout-sidebar">
     <TenancyMenu></TenancyMenu>
